discord install needs sudo perms

+++ discord_from_tar.sh
@@ -4,10 +4,15 @@ set -e
 # Install latest stable Discord from tar.gz
 
 # Get tarball
-wget "https://discordapp.com/api/download/stable?platform=linux&format=tar.gz" -O ~/Downloads/discord.tar.gz
+if [[ ! -f ~/Downloads/discord.tar.gz ]]; then
+  wget "https://discordapp.com/api/download/stable?platform=linux&format=tar.gz" -O ~/Downloads/discord.tar.gz
+fi
 
+sudo -s <<EOF
 # extract to /opt
-tar -xvf ~/Downloads/discord.tar.gz -C /opt/
+if [[ ! -d /opt/Discord ]]; then
+  tar -xvf ~/Downloads/discord.tar.gz -C /opt/
+fi
 
 # symlink binary into PATH
 ln -s /opt/Discord/Discord /usr/local/bin/discord
@@ -18,3 +23,8 @@ sed -i 's+Exec=/usr/share/discord/Discord+Exec=/usr/local/bin/discord+g' /opt/Di
 
 # ... and symlink to expose icon in GNOME menus/search
 ln -s /opt/Discord/discord.desktop ~/.local/share/applications
+
+EOF
+
+# cleanup
+rm ~/Downloads/discord.tar.gz
